完善 ImageLab24 功能，使用 OpenCV 的算法实现

+++ trunk/src/Orc.SmartImage.Common/UnmanagedImage/ImageCieXyz.cs
@@ -6,23 +6,23 @@ namespace Orc.SmartImage
 {
     public partial struct CieXyz
     {
-        public static readonly CieXyz D65 = new CieXyz { X = 0.9505, Y = 1.0, Z = 1.0890 };
+        public static readonly CieXyz D65 = new CieXyz { X = 0.9505f, Y = 1.0f, Z = 1.0890f };
 
-        public double X { get; set; }
-        public double Y { get; set; }
-        public double Z { get; set; }
+        public float X { get; set; }
+        public float Y { get; set; }
+        public float Z { get; set; }
 
         public CieLab ToCieLab()
         {
-            double l = 116.0 * Transform(Y / D65.Y) - 16;
-            double a = 500.0 * (Transform(X / D65.X) - Transform(Y / D65.Y));
-            double b = 200.0 * (Transform(Y / D65.Y) - Transform(Z / D65.Z));
+            float l = 116.0f * Transform(Y / D65.Y) - 16;
+            float a = 500.0f * (Transform(X / D65.X) - Transform(Y / D65.Y));
+            float b = 200.0f * (Transform(Y / D65.Y) - Transform(Z / D65.Z));
             return new CieLab { L = l, A = a, B = b };
         }
 
-        private static double Transform(double t)
+        private static float Transform(float t)
         {
-            return ((t > 0.008856) ? Math.Pow(t, (1.0 / 3.0)) : (7.787 * t + 16.0 / 116.0));
+            return ((t > 0.008856) ? (float) Math.Pow(t, (1.0f / 3.0f)) : (7.787f * t + 16.0f / 116.0f));
         }
     }
 }
